+++ .gitignore
@@ -21,3 +21,4 @@ yarn-debug.log*
 yarn-error.log*
 
 .idea/workspace.xml
+.idea/inspectionProfiles/Project_Default.xml
