+++ src/components/App/App.jsx
@@ -1,119 +1,18 @@
 import React, { Component } from "react";
-import PropTypes from "prop-types";
+import { compose } from 'recompose';
 
-import { withStyles } from "material-ui/styles";
+import { withWidth } from 'material-ui';
+import { withStyles } from 'material-ui/styles';
+import Divider from 'material-ui/Divider';
+import TextField from 'material-ui/TextField';
+import { CircularProgress } from 'material-ui/Progress';
+import Paper from 'material-ui/Paper';
+import Grid from 'material-ui/Grid';
 
-import Paper from "material-ui/Paper";
-import Grid from "material-ui/Grid";
-import Button from "material-ui/Button";
-
-import { MenuItem } from "material-ui/Menu";
-import List, {
-  ListItem,
-  ListItemSecondaryAction,
-  ListItemText
-} from 'material-ui/List';
-
-import Avatar from "material-ui/Avatar";
-import Divider from "material-ui/Divider";
-import TextField from "material-ui/TextField";
-import { CircularProgress } from "material-ui/Progress";
-
-import Tooltip from "material-ui/Tooltip";
-
-import match from "autosuggest-highlight/match";
-import parse from "autosuggest-highlight/parse";
-
-import "./App.css";
-import ModalDetailsAluno from "../ModalDetailsAluno";
-import { compose } from "recompose";
-import { withWidth } from "material-ui";
-import capitalize from "lodash/capitalize";
-import { isWidthDown, isWidthUp } from "material-ui/utils/withWidth";
-
-const styles = theme => ({
-  root: {
-    width: "100%",
-    maxWidth: 360,
-    background: theme.palette.background.paper
-  },
-
-  // Autocomplete
-  suggestionsContainerOpen: {
-    position: "absolute",
-    marginTop: theme.spacing.unit,
-    marginBottom: theme.spacing.unit * 3,
-    left: 0,
-    right: 0
-  },
-  suggestion: {
-    display: "block"
-  },
-  suggestionsList: {
-    margin: 0,
-    padding: 0,
-    listStyleType: "none"
-  },
-  textField: {
-    width: "100%"
-  }
-});
-
-function renderSuggestion(suggestion, { query, isHighlighted }) {
-  const matches = match(suggestion.label, query);
-  const parts = parse(suggestion.label, matches);
-
-  return (
-    <MenuItem selected={isHighlighted} component="div">
-      <div>
-        {parts.map((part, index) => {
-          return part.highlight ? (
-            <span key={index} style={{ fontWeight: 300 }}>
-              {part.text}
-            </span>
-          ) : (
-            <strong key={index} style={{ fontWeight: 500 }}>
-              {part.text}
-            </strong>
-          );
-        })}
-      </div>
-    </MenuItem>
-  );
-}
-
-function renderSuggestionsContainer(options) {
-  const { containerProps, children } = options;
-
-  return (
-    <Paper {...containerProps} square>
-      {children}
-    </Paper>
-  );
-}
-
-const InputSearchAluno = inputProps => {
-  const { classes, autoFocus, value, ref, ...other } = inputProps;
-
-  return (
-    <TextField
-      autoFocus={autoFocus}
-      className={classes.textField}
-      value={value}
-      inputRef={ref}
-      InputProps={{
-        classes: {
-          input: classes.input
-        },
-        ...other
-      }}
-    />
-  );
-};
-const capitalizeEvery = (...words) =>
-  words.map(word => capitalize(word)).join(" ");
-
-console.log("CAPIRALIZE", capitalizeEvery("rod", "costa", "clemente"));
+import './App.css';
+import ListAluno from '../ListAluno';
+import ModalDetailsAluno from '../ModalDetailsAluno';
+import globalStyles from './globalStyles';
 
 /**
  * Start with the slowest value as low end devices often have a small screen.
@@ -208,51 +107,12 @@ class App extends Component {
 
           <Paper>
             {alunos ? (
-              <List {...isWidthDown("md", width) && { dense: true }}>
-                {alunos.map((aluno, index) => (
-                  <Tooltip
-                    id="tooltip-icon"
-                    title={`Ver detalhes sobre: ${aluno.name.first}`}
-                    enterDelay={300}
-                    placement="bottom"
-                    key={index}
-                  >
-                    <ListItem
-                      key={index}
-                      className={classes.listItem}
-                      divider={true}
-                    >
-                      {/*Current width: {width}*/}
-                      <Avatar alt="Remy Sharp" src={aluno.picture.large}/>
-                      <ListItemText
-                        primary={capitalizeEvery(
-                          aluno.name.first,
-                          aluno.name.last
-                        )}
-                        {...isWidthUp("sm", width) && {
-                          secondary: aluno.email
-                        }}
-                      />
-                      <ListItemSecondaryAction>
-                        <Button
-                          className={classes.button}
-                          dense
-                          onClick={() => {
-                            this.handleClickOpen(aluno, index);
-                          }}
-                        >
-                          {isWidthUp("sm", width) && "Ver detalhes"}
-                        </Button>
-
-                        {/*<Checkbox*/}
-                        {/*onChange={this.handleToggle(index)}*/}
-                        {/*checked={this.state.checked.indexOf(index) !== -1}*/}
-                        {/*/>*/}
-                      </ListItemSecondaryAction>
-                    </ListItem>
-                  </Tooltip>
-                ))}
-              </List>
+              <ListAluno
+                alunos={alunos}
+                width={width}
+                classes={classes}
+                handleClickOpen={this.handleClickOpen}
+              />
             ) : (
               <CircularProgress />
             )}
@@ -265,4 +125,4 @@ class App extends Component {
   }
 }
 
-export default compose(withWidth(), withStyles(styles))(App);
+export default compose(withWidth(), withStyles(globalStyles))(App);
