+++ src/components/InputSearchAluno/InputSearchAluno.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import TextField from 'material-ui/TextField';
+
+const InputSearchAluno = inputProps => {
+  const { classes, autoFocus, value, ref, ...other } = inputProps;
+
+  return (
+    <TextField
+      autoFocus={autoFocus}
+      className={classes.textField}
+      value={value}
+      inputRef={ref}
+      InputProps={{
+        classes: {
+          input: classes.input
+        },
+        ...other
+      }}
+    />
+  );
+};
+
+export default InputSearchAluno;
