+++ src/components/InputSearchAluno/index.js
@@ -1,0 +1,2 @@
+import InputSearchAluno from './InputSearchAluno';
+export default InputSearchAluno;
