+++ src/components/ListAluno/index.js
@@ -1,0 +1,2 @@
+import ListAluno from './ListAluno';
+export default ListAluno;
