+++ src/components/ModalDetailsAluno/index.js
@@ -1,6 +1,2 @@
 import ModalDetailsAluno from './ModalDetailsAluno';
 export default ModalDetailsAluno;
-
-// export {default as ModalDetailsAluno} from './ModalDetailsAluno';
-
-// export {default as ModalDetailsAluno} from './ModalDetailsAluno';
