+++ src/components/WebDevApp/index.js
@@ -1,0 +1,2 @@
+import WebDevApp from './WebDevApp';
+export default WebDevApp;
