+++ src/index.js
@@ -2,18 +2,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/App/App';
 import registerServiceWorker from './registerServiceWorker';
-import { MuiThemeProvider } from 'material-ui/styles';
+import WebDevApp from './components/WebDevApp';
 
-import theme from './theme';
-
-const MaterialApp = () => (
-  <MuiThemeProvider theme={theme}>
-    <App />
-  </MuiThemeProvider>
-);
-
-ReactDOM.render(<MaterialApp />, document.getElementById("root"));
+ReactDOM.render(<WebDevApp />, document.getElementById("root"));
 
 registerServiceWorker();
