+++ src/utils/capitalizeEvery.js
@@ -1,0 +1,5 @@
+import capitalize from 'lodash/capitalize';
+
+export default function capitalizeEvery(...words) {
+  return words.map(word => capitalize(word)).join(" ");
+}
